--- src/test/java/com/grahamedgecombe/jterminal/AbstractTerminalModel_ESTest.java
@@ -1,209 +1,0 @@
-/*
- * This file was automatically generated by EvoSuite
- * Fri Jun 30 15:27:56 GMT 2017
- */
-
-package com.grahamedgecombe.jterminal;
-
-import org.junit.Test;
-import static org.junit.Assert.*;
-import static org.evosuite.runtime.EvoAssertions.*;
-import com.grahamedgecombe.jterminal.TerminalCell;
-import com.grahamedgecombe.jterminal.bell.NopBellStrategy;
-import com.grahamedgecombe.jterminal.vt100.Vt100TerminalModel;
-import org.evosuite.runtime.EvoRunner;
-import org.evosuite.runtime.EvoRunnerParameters;
-import org.junit.runner.RunWith;
-
-@RunWith(EvoRunner.class) @EvoRunnerParameters(mockJVMNonDeterminism = true, useVFS = true, useVNET = true, resetStaticState = true, separateClassLoader = true, useJEE = true) 
-public class AbstractTerminalModel_ESTest extends AbstractTerminalModel_ESTest_scaffolding {
-
-  @Test(timeout = 4000)
-  public void test0()  throws Throwable  {
-      Vt100TerminalModel vt100TerminalModel0 = new Vt100TerminalModel(742, 0, 1850);
-      vt100TerminalModel0.print("R{");
-      vt100TerminalModel0.clear();
-      vt100TerminalModel0.getBufferSize();
-      vt100TerminalModel0.getCursorColumn();
-      vt100TerminalModel0.moveCursorUp(0);
-      vt100TerminalModel0.clear();
-      vt100TerminalModel0.getBufferSize();
-      vt100TerminalModel0.clear();
-      vt100TerminalModel0.moveCursorUp(0);
-      vt100TerminalModel0.moveCursorDown(65);
-      vt100TerminalModel0.moveCursorUp(0);
-      // Undeclared exception!
-      try { 
-        vt100TerminalModel0.moveCursorDown((-1903));
-        fail("Expecting exception: IllegalArgumentException");
-      
-      } catch(IllegalArgumentException e) {
-         //
-         // n must be positive
-         //
-         verifyException("com.grahamedgecombe.jterminal.AbstractTerminalModel", e);
-      }
-  }
-
-  @Test(timeout = 4000)
-  public void test1()  throws Throwable  {
-      Vt100TerminalModel vt100TerminalModel0 = new Vt100TerminalModel(4740, 4740);
-      vt100TerminalModel0.getCursorRow();
-      vt100TerminalModel0.getColumns();
-      // Undeclared exception!
-      vt100TerminalModel0.clear();
-  }
-
-  @Test(timeout = 4000)
-  public void test2()  throws Throwable  {
-      Vt100TerminalModel vt100TerminalModel0 = new Vt100TerminalModel();
-      vt100TerminalModel0.clear();
-      NopBellStrategy nopBellStrategy0 = new NopBellStrategy();
-      vt100TerminalModel0.setBellStrategy(nopBellStrategy0);
-      vt100TerminalModel0.getCell(0, 2);
-      vt100TerminalModel0.print("G+'Azq`~qkF\bN");
-      vt100TerminalModel0.getCursorColumn();
-      vt100TerminalModel0.moveCursorForward(0);
-      // Undeclared exception!
-      try { 
-        vt100TerminalModel0.moveCursorBack((-5292));
-        fail("Expecting exception: IllegalArgumentException");
-      
-      } catch(IllegalArgumentException e) {
-         //
-         // n must be positive
-         //
-         verifyException("com.grahamedgecombe.jterminal.AbstractTerminalModel", e);
-      }
-  }
-
-  @Test(timeout = 4000)
-  public void test3()  throws Throwable  {
-      Vt100TerminalModel vt100TerminalModel0 = new Vt100TerminalModel(90, 90, 90);
-      vt100TerminalModel0.moveCursorDown(90);
-      vt100TerminalModel0.moveCursorDown(0);
-      // Undeclared exception!
-      try { 
-        vt100TerminalModel0.getCell(90, (-1577));
-        fail("Expecting exception: IndexOutOfBoundsException");
-      
-      } catch(IndexOutOfBoundsException e) {
-         //
-         // no message in exception (getMessage() returned null)
-         //
-         verifyException("com.grahamedgecombe.jterminal.vt100.Vt100TerminalModel", e);
-      }
-  }
-
-  @Test(timeout = 4000)
-  public void test4()  throws Throwable  {
-      Vt100TerminalModel vt100TerminalModel0 = new Vt100TerminalModel();
-      vt100TerminalModel0.moveCursorDown(0);
-      vt100TerminalModel0.print("");
-      vt100TerminalModel0.getCell(0, 0);
-      vt100TerminalModel0.setCell(0, 0, (TerminalCell) null);
-      vt100TerminalModel0.moveCursorForward(2583);
-      vt100TerminalModel0.clear();
-      vt100TerminalModel0.moveCursorBack(0);
-      vt100TerminalModel0.print("");
-      assertEquals(79, vt100TerminalModel0.getCursorColumn());
-  }
-
-  @Test(timeout = 4000)
-  public void test5()  throws Throwable  {
-      Vt100TerminalModel vt100TerminalModel0 = new Vt100TerminalModel();
-      // Undeclared exception!
-      try { 
-        vt100TerminalModel0.moveCursorForward((-507));
-        fail("Expecting exception: IllegalArgumentException");
-      
-      } catch(IllegalArgumentException e) {
-         //
-         // n must be positive
-         //
-         verifyException("com.grahamedgecombe.jterminal.AbstractTerminalModel", e);
-      }
-  }
-
-  @Test(timeout = 4000)
-  public void test6()  throws Throwable  {
-      Vt100TerminalModel vt100TerminalModel0 = new Vt100TerminalModel();
-      int int0 = 0;
-      vt100TerminalModel0.moveCursorBack(0);
-      int int1 = (-1446);
-      // Undeclared exception!
-      try { 
-        vt100TerminalModel0.moveCursorDown((-1446));
-        fail("Expecting exception: IllegalArgumentException");
-      
-      } catch(IllegalArgumentException e) {
-         //
-         // n must be positive
-         //
-         verifyException("com.grahamedgecombe.jterminal.AbstractTerminalModel", e);
-      }
-  }
-
-  @Test(timeout = 4000)
-  public void test7()  throws Throwable  {
-      Vt100TerminalModel vt100TerminalModel0 = new Vt100TerminalModel();
-      vt100TerminalModel0.moveCursorBack(65);
-      vt100TerminalModel0.moveCursorDown(65);
-      vt100TerminalModel0.getBufferSize();
-      vt100TerminalModel0.getColumns();
-      vt100TerminalModel0.clear();
-      vt100TerminalModel0.moveCursorUp(25);
-      vt100TerminalModel0.moveCursorDown(97);
-      vt100TerminalModel0.getCursorColumn();
-      vt100TerminalModel0.getCursorRow();
-      // Undeclared exception!
-      try { 
-        vt100TerminalModel0.setCursorColumn((-1));
-        fail("Expecting exception: IllegalArgumentException");
-      
-      } catch(IllegalArgumentException e) {
-         //
-         // column out of range
-         //
-         verifyException("com.grahamedgecombe.jterminal.vt100.Vt100TerminalModel", e);
-      }
-  }
-
-  @Test(timeout = 4000)
-  public void test8()  throws Throwable  {
-      Vt100TerminalModel vt100TerminalModel0 = new Vt100TerminalModel();
-      vt100TerminalModel0.moveCursorUp(0);
-      vt100TerminalModel0.moveCursorDown(0);
-      vt100TerminalModel0.getDefaultBackgroundColor();
-      // Undeclared exception!
-      try { 
-        vt100TerminalModel0.moveCursorBack((-2707));
-        fail("Expecting exception: IllegalArgumentException");
-      
-      } catch(IllegalArgumentException e) {
-         //
-         // n must be positive
-         //
-         verifyException("com.grahamedgecombe.jterminal.AbstractTerminalModel", e);
-      }
-  }
-
-  @Test(timeout = 4000)
-  public void test9()  throws Throwable  {
-      Vt100TerminalModel vt100TerminalModel0 = new Vt100TerminalModel(7, 7);
-      int int0 = 0;
-      vt100TerminalModel0.moveCursorForward(0);
-      vt100TerminalModel0.moveCursorForward(7);
-      // Undeclared exception!
-      try { 
-        vt100TerminalModel0.moveCursorUp((-2021));
-        fail("Expecting exception: IllegalArgumentException");
-      
-      } catch(IllegalArgumentException e) {
-         //
-         // n must be positive
-         //
-         verifyException("com.grahamedgecombe.jterminal.AbstractTerminalModel", e);
-      }
-  }
-}
