--- src/test/java/com/grahamedgecombe/jterminal/JTerminal_ESTest.java
@@ -1,159 +1,0 @@
-/*
- * This file was automatically generated by EvoSuite
- * Fri Jun 30 15:27:48 GMT 2017
- */
-
-package com.grahamedgecombe.jterminal;
-
-import org.junit.Test;
-import static org.junit.Assert.*;
-import static org.evosuite.runtime.EvoAssertions.*;
-import com.grahamedgecombe.jterminal.JTerminal;
-import com.grahamedgecombe.jterminal.TerminalModel;
-import com.grahamedgecombe.jterminal.vt100.Vt100TerminalModel;
-import java.awt.Dimension;
-import org.evosuite.runtime.EvoRunner;
-import org.evosuite.runtime.EvoRunnerParameters;
-import org.junit.runner.RunWith;
-
-@RunWith(EvoRunner.class) @EvoRunnerParameters(mockJVMNonDeterminism = true, useVFS = true, useVNET = true, resetStaticState = true, separateClassLoader = true, useJEE = true) 
-public class JTerminal_ESTest extends JTerminal_ESTest_scaffolding {
-
-  @Test(timeout = 4000)
-  public void test00()  throws Throwable  {
-      Vt100TerminalModel vt100TerminalModel0 = new Vt100TerminalModel(0, 0, 0);
-      JTerminal jTerminal0 = new JTerminal(vt100TerminalModel0);
-      TerminalModel terminalModel0 = jTerminal0.getModel();
-      assertSame(vt100TerminalModel0, terminalModel0);
-  }
-
-  @Test(timeout = 4000)
-  public void test01()  throws Throwable  {
-      JTerminal jTerminal0 = new JTerminal();
-      // Undeclared exception!
-      try { 
-        jTerminal0.print((String) null);
-        fail("Expecting exception: NullPointerException");
-      
-      } catch(NullPointerException e) {
-         //
-         // str
-         //
-         verifyException("com.grahamedgecombe.jterminal.vt100.Vt100TerminalModel", e);
-      }
-  }
-
-  @Test(timeout = 4000)
-  public void test02()  throws Throwable  {
-      JTerminal jTerminal0 = new JTerminal();
-      Vt100TerminalModel vt100TerminalModel0 = new Vt100TerminalModel(0, 12);
-      jTerminal0.setModel(vt100TerminalModel0);
-      // Undeclared exception!
-      try { 
-        jTerminal0.print("=Zwqw.rGt");
-        fail("Expecting exception: ArrayIndexOutOfBoundsException");
-      
-      } catch(ArrayIndexOutOfBoundsException e) {
-         //
-         // 0
-         //
-         verifyException("com.grahamedgecombe.jterminal.vt100.Vt100TerminalModel$Vt100Listener", e);
-      }
-  }
-
-  @Test(timeout = 4000)
-  public void test03()  throws Throwable  {
-      JTerminal jTerminal0 = null;
-      try {
-        jTerminal0 = new JTerminal((TerminalModel) null);
-        fail("Expecting exception: NullPointerException");
-      
-      } catch(NullPointerException e) {
-         //
-         // model
-         //
-         verifyException("com.grahamedgecombe.jterminal.JTerminal", e);
-      }
-  }
-
-  @Test(timeout = 4000)
-  public void test04()  throws Throwable  {
-      JTerminal jTerminal0 = new JTerminal();
-      Vt100TerminalModel vt100TerminalModel0 = new Vt100TerminalModel(12, 0, 0);
-      jTerminal0.setModel(vt100TerminalModel0);
-      // Undeclared exception!
-      try { 
-        jTerminal0.println("com.grahamedgecombe.jterminal.JTerminal$Terminal");
-        fail("Expecting exception: ArrayIndexOutOfBoundsException");
-      
-      } catch(ArrayIndexOutOfBoundsException e) {
-         //
-         // -1
-         //
-         verifyException("com.grahamedgecombe.jterminal.vt100.Vt100TerminalModel$Vt100Listener", e);
-      }
-  }
-
-  @Test(timeout = 4000)
-  public void test05()  throws Throwable  {
-      JTerminal jTerminal0 = new JTerminal();
-      jTerminal0.print("com.grahamedgecombe.jterminal.JTerminal");
-      TerminalModel terminalModel0 = jTerminal0.getModel();
-      assertEquals(39, terminalModel0.getCursorColumn());
-  }
-
-  @Test(timeout = 4000)
-  public void test06()  throws Throwable  {
-      JTerminal jTerminal0 = new JTerminal();
-      // Undeclared exception!
-      try { 
-        jTerminal0.println((String) null);
-        fail("Expecting exception: NullPointerException");
-      
-      } catch(NullPointerException e) {
-         //
-         // str
-         //
-         verifyException("com.grahamedgecombe.jterminal.JTerminal", e);
-      }
-  }
-
-  @Test(timeout = 4000)
-  public void test07()  throws Throwable  {
-      JTerminal jTerminal0 = new JTerminal();
-      // Undeclared exception!
-      try { 
-        jTerminal0.setModel((TerminalModel) null);
-        fail("Expecting exception: NullPointerException");
-      
-      } catch(NullPointerException e) {
-         //
-         // model
-         //
-         verifyException("com.grahamedgecombe.jterminal.JTerminal", e);
-      }
-  }
-
-  @Test(timeout = 4000)
-  public void test08()  throws Throwable  {
-      JTerminal jTerminal0 = new JTerminal();
-      jTerminal0.println("");
-      TerminalModel terminalModel0 = jTerminal0.getModel();
-      assertEquals(1, terminalModel0.getCursorRow());
-  }
-
-  @Test(timeout = 4000)
-  public void test09()  throws Throwable  {
-      JTerminal jTerminal0 = new JTerminal();
-      Dimension dimension0 = jTerminal0.getPreferredSize();
-      assertEquals(300, dimension0.height);
-      assertEquals(640, dimension0.width);
-  }
-
-  @Test(timeout = 4000)
-  public void test10()  throws Throwable  {
-      Vt100TerminalModel vt100TerminalModel0 = new Vt100TerminalModel(4517, 521, 4517);
-      JTerminal jTerminal0 = new JTerminal(vt100TerminalModel0);
-      assertFalse(jTerminal0.isFocusTraversalPolicySet());
-  }
-}
