--- src/test/java/com/grahamedgecombe/jterminal/bell/NopBellStrategy_ESTest.java
@@ -1,23 +1,0 @@
-/*
- * This file was automatically generated by EvoSuite
- * Fri Jun 30 15:29:21 GMT 2017
- */
-
-package com.grahamedgecombe.jterminal.bell;
-
-import org.junit.Test;
-import static org.junit.Assert.*;
-import com.grahamedgecombe.jterminal.bell.NopBellStrategy;
-import org.evosuite.runtime.EvoRunner;
-import org.evosuite.runtime.EvoRunnerParameters;
-import org.junit.runner.RunWith;
-
-@RunWith(EvoRunner.class) @EvoRunnerParameters(mockJVMNonDeterminism = true, useVFS = true, useVNET = true, resetStaticState = true, separateClassLoader = true, useJEE = true) 
-public class NopBellStrategy_ESTest extends NopBellStrategy_ESTest_scaffolding {
-
-  @Test(timeout = 4000)
-  public void test0()  throws Throwable  {
-      NopBellStrategy nopBellStrategy0 = new NopBellStrategy();
-      nopBellStrategy0.soundBell();
-  }
-}
