--- src/test/java/com/grahamedgecombe/jterminal/vt100/AnsiControlSequence_ESTest.java
@@ -1,95 +1,0 @@
-/*
- * This file was automatically generated by EvoSuite
- * Fri Jun 30 15:29:05 GMT 2017
- */
-
-package com.grahamedgecombe.jterminal.vt100;
-
-import org.junit.Test;
-import static org.junit.Assert.*;
-import static org.evosuite.runtime.EvoAssertions.*;
-import com.grahamedgecombe.jterminal.vt100.AnsiControlSequence;
-import org.evosuite.runtime.EvoRunner;
-import org.evosuite.runtime.EvoRunnerParameters;
-import org.junit.runner.RunWith;
-
-@RunWith(EvoRunner.class) @EvoRunnerParameters(mockJVMNonDeterminism = true, useVFS = true, useVNET = true, resetStaticState = true, separateClassLoader = true, useJEE = true) 
-public class AnsiControlSequence_ESTest extends AnsiControlSequence_ESTest_scaffolding {
-
-  @Test(timeout = 4000)
-  public void test0()  throws Throwable  {
-      String[] stringArray0 = new String[1];
-      stringArray0[0] = "fxV<|9*0{9P<";
-      AnsiControlSequence ansiControlSequence0 = new AnsiControlSequence('\"', stringArray0);
-      char char0 = ansiControlSequence0.getCommand();
-      assertEquals('\"', char0);
-  }
-
-  @Test(timeout = 4000)
-  public void test1()  throws Throwable  {
-      String[] stringArray0 = new String[0];
-      AnsiControlSequence ansiControlSequence0 = new AnsiControlSequence('0', stringArray0);
-      char char0 = ansiControlSequence0.getCommand();
-      assertEquals('0', char0);
-  }
-
-  @Test(timeout = 4000)
-  public void test2()  throws Throwable  {
-      String[] stringArray0 = new String[1];
-      stringArray0[0] = "";
-      AnsiControlSequence ansiControlSequence0 = new AnsiControlSequence('w', stringArray0);
-      assertEquals('w', ansiControlSequence0.getCommand());
-  }
-
-  @Test(timeout = 4000)
-  public void test3()  throws Throwable  {
-      String[] stringArray0 = new String[1];
-      stringArray0[0] = "y_n$";
-      AnsiControlSequence ansiControlSequence0 = new AnsiControlSequence('y', stringArray0);
-      ansiControlSequence0.getParameters();
-      assertEquals('y', ansiControlSequence0.getCommand());
-  }
-
-  @Test(timeout = 4000)
-  public void test4()  throws Throwable  {
-      String[] stringArray0 = new String[1];
-      AnsiControlSequence ansiControlSequence0 = null;
-      try {
-        ansiControlSequence0 = new AnsiControlSequence('4', stringArray0);
-        fail("Expecting exception: NullPointerException");
-      
-      } catch(NullPointerException e) {
-      }
-  }
-
-  @Test(timeout = 4000)
-  public void test5()  throws Throwable  {
-      AnsiControlSequence ansiControlSequence0 = null;
-      try {
-        ansiControlSequence0 = new AnsiControlSequence('D', (String[]) null);
-        fail("Expecting exception: NullPointerException");
-      
-      } catch(NullPointerException e) {
-         //
-         // parameters
-         //
-         verifyException("com.grahamedgecombe.jterminal.vt100.AnsiControlSequence", e);
-      }
-  }
-
-  @Test(timeout = 4000)
-  public void test6()  throws Throwable  {
-      String[] stringArray0 = new String[0];
-      AnsiControlSequence ansiControlSequence0 = new AnsiControlSequence('B', stringArray0);
-      ansiControlSequence0.getParameters();
-      assertEquals('B', ansiControlSequence0.getCommand());
-  }
-
-  @Test(timeout = 4000)
-  public void test7()  throws Throwable  {
-      String[] stringArray0 = new String[0];
-      AnsiControlSequence ansiControlSequence0 = new AnsiControlSequence('B', stringArray0);
-      char char0 = ansiControlSequence0.getCommand();
-      assertEquals('B', char0);
-  }
-}
