--- src/test/java/com/grahamedgecombe/jterminal/vt100/Vt100TerminalModel_ESTest_scaffolding.java
@@ -1,115 +1,0 @@
-/**
- * Scaffolding file used to store all the setups needed to run 
- * tests automatically generated by EvoSuite
- * Fri Jun 30 15:26:16 GMT 2017
- */
-
-package com.grahamedgecombe.jterminal.vt100;
-
-import org.evosuite.runtime.annotation.EvoSuiteClassExclude;
-import org.junit.BeforeClass;
-import org.junit.Before;
-import org.junit.After;
-import org.junit.AfterClass;
-import org.evosuite.runtime.sandbox.Sandbox;
-import org.evosuite.runtime.sandbox.Sandbox.SandboxMode;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
-import java.util.concurrent.TimeUnit;
-
-@EvoSuiteClassExclude
-public class Vt100TerminalModel_ESTest_scaffolding {
-
-  @org.junit.Rule 
-  public org.evosuite.runtime.vnet.NonFunctionalRequirementRule nfr = new org.evosuite.runtime.vnet.NonFunctionalRequirementRule();
-
-  protected static ExecutorService executor; 
-
-  private static final java.util.Properties defaultProperties = (java.util.Properties) java.lang.System.getProperties().clone(); 
-
-  private org.evosuite.runtime.thread.ThreadStopper threadStopper =  new org.evosuite.runtime.thread.ThreadStopper (org.evosuite.runtime.thread.KillSwitchHandler.getInstance(), 3000);
-
-
-  @BeforeClass 
-  public static void initEvoSuiteFramework() { 
-    org.evosuite.runtime.RuntimeSettings.className = "com.grahamedgecombe.jterminal.vt100.Vt100TerminalModel"; 
-    org.evosuite.runtime.GuiSupport.initialize(); 
-    org.evosuite.runtime.RuntimeSettings.maxNumberOfThreads = 100; 
-    org.evosuite.runtime.RuntimeSettings.maxNumberOfIterationsPerLoop = 10000; 
-    org.evosuite.runtime.RuntimeSettings.mockSystemIn = true; 
-    org.evosuite.runtime.RuntimeSettings.sandboxMode = org.evosuite.runtime.sandbox.Sandbox.SandboxMode.RECOMMENDED; 
-    org.evosuite.runtime.sandbox.Sandbox.initializeSecurityManagerForSUT(); 
-    executor = Executors.newCachedThreadPool(); 
-    org.evosuite.runtime.classhandling.JDKClassResetter.init();
-    setSystemProperties();
-    initializeClasses();
-    org.evosuite.runtime.Runtime.getInstance().resetRuntime(); 
-  } 
-
-  @AfterClass 
-  public static void clearEvoSuiteFramework(){ 
-    Sandbox.resetDefaultSecurityManager(); 
-    executor.shutdownNow(); 
-    java.lang.System.setProperties((java.util.Properties) defaultProperties.clone()); 
-  } 
-
-  @Before 
-  public void initTestCase(){ 
-    threadStopper.storeCurrentThreads();
-    threadStopper.startRecordingTime();
-    org.evosuite.runtime.jvm.ShutdownHookHandler.getInstance().initHandler(); 
-    org.evosuite.runtime.sandbox.Sandbox.goingToExecuteSUTCode(); 
-    setSystemProperties(); 
-    org.evosuite.runtime.GuiSupport.setHeadless(); 
-    org.evosuite.runtime.Runtime.getInstance().resetRuntime(); 
-    org.evosuite.runtime.agent.InstrumentingAgent.activate(); 
-  } 
-
-  @After 
-  public void doneWithTestCase(){ 
-    threadStopper.killAndJoinClientThreads();
-    org.evosuite.runtime.jvm.ShutdownHookHandler.getInstance().safeExecuteAddedHooks(); 
-    org.evosuite.runtime.classhandling.JDKClassResetter.reset(); 
-    resetClasses(); 
-    org.evosuite.runtime.sandbox.Sandbox.doneWithExecutingSUTCode(); 
-    org.evosuite.runtime.agent.InstrumentingAgent.deactivate(); 
-    org.evosuite.runtime.GuiSupport.restoreHeadlessMode(); 
-  } 
-
-  public static void setSystemProperties() {
- 
-    java.lang.System.setProperties((java.util.Properties) defaultProperties.clone()); 
-    java.lang.System.setProperty("file.encoding", "UTF-8"); 
-    java.lang.System.setProperty("java.awt.headless", "true"); 
-    java.lang.System.setProperty("user.country", "US"); 
-    java.lang.System.setProperty("user.language", "en"); 
-    java.lang.System.setProperty("user.timezone", "Europe/Brussels"); 
-  }
-
-  private static void initializeClasses() {
-    org.evosuite.runtime.classhandling.ClassStateSupport.initializeClasses(Vt100TerminalModel_ESTest_scaffolding.class.getClassLoader() ,
-      "com.grahamedgecombe.jterminal.bell.BellStrategy",
-      "com.grahamedgecombe.jterminal.vt100.Vt100TerminalModel",
-      "com.grahamedgecombe.jterminal.vt100.Vt100TerminalModel$Vt100Listener",
-      "com.grahamedgecombe.jterminal.TerminalCell",
-      "com.grahamedgecombe.jterminal.TerminalModel",
-      "com.grahamedgecombe.jterminal.vt100.AnsiControlSequenceListener",
-      "com.grahamedgecombe.jterminal.bell.NopBellStrategy",
-      "com.grahamedgecombe.jterminal.bell.BeepBellStrategy",
-      "com.grahamedgecombe.jterminal.vt100.AnsiControlSequenceParser",
-      "com.grahamedgecombe.jterminal.AbstractTerminalModel",
-      "com.grahamedgecombe.jterminal.vt100.SgrColor"
-    );
-  } 
-
-  private static void resetClasses() {
-    org.evosuite.runtime.classhandling.ClassResetter.getInstance().setClassLoader(Vt100TerminalModel_ESTest_scaffolding.class.getClassLoader()); 
-
-    org.evosuite.runtime.classhandling.ClassStateSupport.resetClasses(
-      "com.grahamedgecombe.jterminal.vt100.Vt100TerminalModel",
-      "com.grahamedgecombe.jterminal.vt100.SgrColor",
-      "com.grahamedgecombe.jterminal.vt100.AnsiControlSequenceParser"
-    );
-  }
-}
